+++ cmd/version/git_native_disabled.go
@@ -1,0 +1,7 @@
+//go:build !gitenabled
+
+package version
+
+func isGitEnabled() bool {
+	return false
+}
